Encapsulated some of the linear algebra for Tables/AlignedVectors
